requirements.txt
numpy==1.25.2
pandas==2.1.0
python-binance==1.0.17
python-dotenv==1.0.0 